requirements.txt
yfinance
pandas
matplotlib
pymongo
chromadb
sentence-transformers
python-dotenv
flask==2.3.2
gunicorn==20.1.0
pymongo==4.3.3
python-dotenv==1.0.0
pandas==2.0.3